added formatter hook
